Fixing Building Scripts for LibUSB

+++ build_scripts/build_app_deps.sh
@@ -40,25 +40,38 @@ cp lib/libSoapySDR* $BUILD_DIR/lib/
 cp lib/SoapySDR.pc $BUILD_DIR/lib/pkgconfig/SoapySDR.pc
 cp ../include/SoapySDR/* $BUILD_DIR/include/SoapySDR/
 
-# build libusb
-cd "$BUILD_DIR"
-git clone https://github.com/libusb/libusb.git
-cd libusb
-git reset --hard d52e355daa09f17ce64819122cb067b8a2ee0d4b
-./autogen.sh
-./configure
-make
-cp ./libusb/.libs/libusb*$LIB_EXT $BUILD_DIR/lib
-cp ./libusb-1.0.pc $BUILD_DIR/lib/pkgconfig/libusb-1.0.pc
-sed -i '' "1s|.*|prefix=$BUILD_DIR|" $BUILD_DIR/lib/pkgconfig/libusb-1.0.pc
+# build libusb if libs don't already exist
+if [ ! -f "$BUILD_DIR/lib/libusb-1.0$LIB_EXT" ]; then
+  cd "$BUILD_DIR"
+  git clone https://github.com/libusb/libusb.git
+  cd libusb
+  git reset --hard d52e355daa09f17ce64819122cb067b8a2ee0d4b
+  ./autogen.sh
+  ./configure
+  make
+  mkdir $BUILD_DIR/include/libusb-1.0
+  cp ./libusb/libusb.h $BUILD_DIR/include/libusb-1.0/libusb.h
+  cp ./libusb/.libs/libusb*$LIB_EXT $BUILD_DIR/lib
+  cp ./libusb-1.0.pc $BUILD_DIR/lib/pkgconfig/libusb-1.0.pc
+  sed -i '' "1s|.*|prefix=$BUILD_DIR|" $BUILD_DIR/lib/pkgconfig/libusb-1.0.pc
+fi
 
 # build librtlsdr
 cd "$BUILD_DIR"
 git clone https://github.com/osmocom/rtl-sdr.git
 cd rtl-sdr
+git reset --hard d52e355daa09f17ce64819122cb067b8a2ee0d4b
+# patch cmake file with custom libusb installation path
+git checkout -- CMakeLists.txt
+PATCH_DATA="/if(PKG_CONFIG_FOUND AND NOT LIBUSB_FOUND)/i\\
+set(LIBUSB_LIBRARIES \"$BUILD_DIR/lib/libusb-1.0.0$LIB_EXT\")\\
+set(LIBUSB_INCLUDE_DIRS \"$BUILD_DIR/include/libusb-1.0\")\\
+set(LIBUSB_FOUND TRUE)\\
+"
+sed -i '' "$PATCH_DATA" CMakeLists.txt
 mkdir build
 cd build
-cmake -DCMAKE_PREFIX_PATH=$BUILD_DIR/lib ../
+cmake ../
 make
 cp src/librtlsdr* $BUILD_DIR/lib
 cp ../include/{rtl-sdr.h,rtl-sdr_export.h} $BUILD_DIR/include
